# express-back/docs/swagger.generated.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "哈尔滨学院校园综合服务平台 API",
    "version": "1.0.0",
    "description": "校园综合服务平台后端接口文档。"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Current server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/admin/reviews/workbench": {
      "get": {
        "tags": [
          "Admin Reviews"
        ],
        "summary": "Get Workbench",
        "operationId": "ReviewController_getWorkbench",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server Error"
          }
        },
        "x-auto-generated": true,
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/reviews/workbench/reports/{id}": {
      "patch": {
        "tags": [
          "Admin Reviews"
        ],
        "summary": "Handle Report",
        "operationId": "ReviewController_handleReport",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server Error"
          }
        },
        "x-auto-generated": true,
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  }
}